requirements.txt
Django==5.0
djangorestframework==3.14.0
drf_yasg==1.21.7
